docs/ideas/05-Claude-Desktop-Research/requirements.txt
# requirements.txt — pinned majors, loose patch for security updates
python-telegram-bot[job-queue]==21.*
pydantic-ai[anthropic]>=1.0,<2
pydantic-settings>=2.0,<3
logfire[pydantic-ai]>=3.0,<4
